pyproject.toml
[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[project]
name        = "xsecure"
version     = "1.0.0"
description = "Autonomous Incident Response & Threat Mitigation — OpenEnv RL environment."
readme      = "README.md"
requires-python = ">=3.10"
license     = { text = "Apache-2.0" }
keywords    = ["reinforcement-learning", "cybersecurity", "openenv", "incident-response"]

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]
